Add feature and fixed bugs
Added negation, symbol ! (ex. !title), that returns true if the entry
value does not exist. This is an easy way to make an if and else
statements.
When going back to the page, if the search bar has words then the
elements are hidden again.
Latex that didn't work properly has been removed for now.