Replaced simulated Greek capital tonos latex encodings
Based on the translation for Ί U+038A to \'{}{I}, I added a set of empty
{} after the \' for several capital greek (with tonos) characters, such
as Ό u+038C and Έ u+0388. Note that these LaTeX commands won't produce
the true unicode character, but rather an apostrophe followed by the
capital letter. Not being an expert in Greek, I'm not sure how likely
these encodings would appear in a BibTeX file.