:wrench::rotating_light: Updates by TestCafe broke tests
The updates by TestCafe seems to have broken the conversion of strings
to number. In this case, the year was no longer being correctly
converted to a number and the test was failing because it expects a
number. I have added an explicit conversion of the string to Number in
the javascript to correct for this issue.