:bug: Could not use #bibtex_input with class="bibtex_display"
There was no need for this if statement as we can combined the
bibtex_input with the bibtex files. The issue was that if bibtex_input
was used then it was not possible to use the class form of
bibtex_display. In particular this became an issue when writing tests
for multiple bibtex_displays and using bibtex_input within the html file.