:sparkles::rotating_light: Setting default author format
The author output will now have a default format to ensure that the
output is normalized for all different variations of entering names
into BibTeX.

Default: First von Last, Junior

Added bibtex-js-rif (Replace If First) option to clear text inside the
span if the firname or von part is missing.

Test: they have been updated so that the author format matches the
default format. This was simply switching the ordering of last name
and first names.