:boom: Removing @ symbol from bibtextypekey
The @ symbol is used by the search methods and as such it should be
part of the bibtextypekey. The @ symbol is just redundant text that
should have been removed in the first place.

This will break previous uses of the academicstyle (wiki will need
updated) and if anyone used the == with the bibtextypekey.

Test "Check adademic style" has been updated to comply with the
change. Test files ifequals and academicstyle have also been updated.