:bug::rotating_light: Missing ID value on custom group header
The custom group header was missing the id value which meant it did
not work with the topics generation. The value has now been added.

There were also no tests for custom group headers which has been added
to test that the custom group header prints the correct value and that
the id value is being printed correctly.