:sparkles: Adding callback functions
For any bibtex_display element the attribute callback can be used to
run javascript code after the completion of bibtex_js. The code has
access to the bibtex_display element using the variable
`bibtex_display`.

For the bibtex_template, the attribute callback can also be used to
run javascript code after each bibtex entry is added to the
webpage. The code has access to the bibtexentry with the variable
`bibtexentry`.

The callback for bibtex_display is particularly important for any user
that needs to run javascript code after bibtex_js has finished
processing and creating the bibtex entries. The call back for the
bibtex_template provides access during the generation of the bibtex
entries and can be useful for writing custom code for
hiding/altering/removing bibtex entries. Both are meant to be open
ended methods for customing bibtex_js within the users website.