:sparkles::bug: Added bibtexjs-css-escape for bibtexVar
In certain cases the value of generated by bibtexVar would be used by
a javascript function to perform some function. However, there are
symbols allowed in bibtex that aren't allowed in javascript without an
escape character before them. There is now an option to escape the
characters when using bibtexVar.

Notably this is used in the FullDemo.html:

```html
<a class="bibtexVar" role="button" data-toggle="collapse"
          href="#bib+BIBTEXKEY+" aria-expanded="false"
	  aria-controls="bib+BIBTEXKEY+" extra="BIBTEXKEY"
          bibtexjs-css-escape> <!-- Shown here! -->
```

:rotating_light: Two simple tests were added to ensure that the
escaping works with the bibtex key and a bibtex entry value.