:bug: Generate List can handle other author format types
The author format can change now so I am using the class attribute of
the author that has a consistent author format to maintain a
consistent generated list.

I've added a test to check that it workes with a different format and
one that uses initials. This is related to issue #53.

There was a bug in the bibtex values since the Jr. was incorrectly
kept with the first name. This makes it part of the first name and not
a junior.