:rotating_light: Adds test for clicking individual authors to select
This is related to issue #13. After reworking the authors display
there is now a way to add the ability to click authors and search on
them (individually) within the html rather than the javascript. I
prefer this solution as it doesn't further complicate the javascript
code.

The solution is rather simple too and the only complicated part is the
BibTexSearcher funtion that gets copied into the onclick
attribute. However, its mostly copy paste and it leaves room for
customization on the user (rather than building the <a> within
javascript which was very limiting).