✨ Adds support for custom field splitting in auto-generated selects
Squashing: Co-Authored-by: Philip Cooksey <pcooksey@andrew.cmu.edu>
Compressing @mhaeussermann keyword filter work into the new
attribute bibtex_split_by for auto-generated selects.

The functionality needed for the non-standard "keywords" bibtex field is
that it should be split up with commas rather than taken as a
whole. This can be accomplished by adding an attribute to the select that
splits the string as @mhaeussermann had done. Adding an array in the
else statement allows up to have both the default and extended split
functionality without adding too much in terms of if statements.

Removing extra html file and combining it with the previously made
auto generate select page. Tests will come in the next commit